+++ test/TaskOption.ts
@@ -203,13 +203,6 @@ describe('TaskOption', () => {
     U.deepStrictEqual(await g('')(), O.none)
   })
 
-  it('chainOptionK', async () => {
-    const f = _.chainOptionK((n: number) => (n > 0 ? O.some(n) : O.none))
-    U.deepStrictEqual(await f(_.some(1))(), O.some(1))
-    U.deepStrictEqual(await f(_.some(-1))(), O.none)
-    U.deepStrictEqual(await f(_.none)(), O.none)
-  })
-
   it('tryCatchK', async () => {
     const f = (n: number) => (n > 0 ? Promise.resolve(n) : Promise.reject(n))
     const g = _.tryCatchK(f)